requirements.txt
boto3==1.34.143
botocore==1.34.143
cryptography==42.0.8
pyOpenSSL==24.1.0
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2